Fix t.compile for Granite20b   (#282)
* Enable t.compile for Granite20B

Granite20B model uses the GPTBigCode model type. In this model,
model object uses model.transformer instead of model.model for the actual
model. This is a pointed fix for this model, in future other models
may need similar change.

* Enable t.compile for Granite20B

Granite20B model uses the GPTBigCode model type. In this model,
model object uses model.transformer instead of model.model for the actual
model. This is a pointed fix for this model, in future other models
may need similar change.

* fixed a comment